refactor: CLIN-642 remove patient search legacy code
