+++ client/src/actions/patient.ts
@@ -5,35 +5,6 @@ import * as actions from './type';
 
 type Action = (...args: any) => { type: keyof typeof actions; payload?: any };
 
-export const autoCompletePatients: Action = (
-  type: string,
-  query: any,
-  page: number,
-  size: number,
-) => ({
-  payload: {
-    page: page || 1,
-    query: query || null,
-    size: size || 25,
-    type: type || 'partial',
-  },
-  type: actions.PATIENT_AUTOCOMPLETE_REQUESTED,
-});
-
-export const autoCompletePatientsSelected: Action = () => ({
-  payload: {},
-  type: actions.PATIENT_AUTOCOMPLETE_SELECTED,
-});
-
-export const searchPatientsByQuery: Action = (query: any, page: number, size: number) => ({
-  payload: {
-    page: page || 1,
-    query: query || null,
-    size: size || 25,
-  },
-  type: actions.PATIENT_SEARCH_REQUESTED,
-});
-
 export const updateServiceRequestStatus: Action = (
   serviceRequestId: string,
   newStatus: string,
@@ -52,12 +23,6 @@ export const getPatientByRamq: Action = (ramq: string) => ({
   type: actions.PATIENT_FETCH_INFO_BY_RAMQ,
 });
 
-export const changeSearchType: Action = (type: string) => ({
-  payload: {
-    type,
-  },
-  type: actions.CHANGE_SEARCH_TYPE_REQUESTED,
-});
 
 type familyActionCb = ((isSuccess: boolean) => Promise<void> | void) | undefined;
