+++ client/src/actions/type.js
@@ -1,8 +1,4 @@
 /* eslint-disable max-len */
-export const START_LOADING_ANIMATION = 'START_LOADING_ANIMATION';
-export const STOP_LOADING_ANIMATION = 'STOP_LOADING_ANIMATION';
-
-export const START_SUBLOADING_ANIMATION = 'START_SUBLOADING_ANIMATION';
 export const STOP_SUBLOADING_ANIMATION = 'STOP_SUBLOADING_ANIMATION';
 
 export const NOTIFICATION = 'NOTIFICATION';
@@ -16,8 +12,6 @@ export const APP_CHANGE_LANGUAGE_SUCCEEDED = 'APP_CHANGE_LANGUAGE_SUCCEEDED';
 export const APP_CHANGE_LANGUAGE_FAILED = 'APP_CHANGE_LANGUAGE_FAILED';
 
 export const ROUTER_NAVIGATION_REQUESTED = 'ROUTER_NAVIGATION_REQUESTED';
-export const ROUTER_NAVIGATION_SUCCEEDED = 'ROUTER_NAVIGATION_SUCCEEDED';
-export const ROUTER_NAVIGATION_FAILED = 'ROUTER_NAVIGATION_FAILED';
 
 export const NAVIGATION_PATIENT_SCREEN_REQUESTED = 'NAVIGATION_PATIENT_SCREEN_REQUESTED';
 export const NAVIGATION_PATIENT_SCREEN_SUCCEEDED = 'NAVIGATION_PATIENT_SCREEN_SUCCEEDED';
@@ -74,16 +68,6 @@ export const USER_PROFILE_UPDATE_REQUESTED = 'USER_PROFILE_UPDATE_REQUESTED';
 export const USER_PROFILE_UPDATE_SUCCEEDED = 'USER_PROFILE_UPDATE_SUCCEEDED';
 export const USER_PROFILE_UPDATE_FAILED = 'USER_PROFILE_UPDATE_FAILED';
 
-export const USER_PROFILE_UPDATE_COLUMNS = 'USER_PROFILE_UPDATE_COLUMNS';
-export const USER_PROFILE_UPDATE_COLUMNS_ORDER = 'USER_PROFILE_UPDATE_COLUMNS_ORDER';
-
-export const USER_PROFILE_UPDATE_COLUMNS_RESET_REQUESTED =
-  'USER_PROFILE_UPDATE_COLUMNS_RESET_REQUESTED';
-export const USER_PROFILE_UPDATE_COLUMNS_RESET_SUCCEEDED =
-  'USER_PROFILE_UPDATE_COLUMNS_RESET_SUCCEEDED';
-export const USER_PROFILE_UPDATE_COLUMNS_RESET_FAILED = 'USER_PROFILE_UPDATE_COLUMNS_RESET_FAILED';
-export const USER_PROFILE_UPDATE_COLUMNS_RESET = 'USER_PROFILE_UPDATE_COLUMNS_RESET';
-
 export const UPDATE_USER_AUTH_PERMISSIONS_REQUESTED = 'UPDATE_USER_AUTH_PERMISSIONS_REQUESTED';
 export const UPDATE_USER_AUTH_PERMISSIONS_SUCCEEDED = 'UPDATE_USER_AUTH_PERMISSIONS_SUCCEEDED';
 export const UPDATE_USER_AUTH_PERMISSIONS_FAILED = 'UPDATE_USER_AUTH_PERMISSIONS_FAILED';
@@ -115,16 +99,7 @@ export const PATIENT_REMOVE_PARENT_SUCCEEDED = 'PATIENT_REMOVE_PARENT_SUCCEEDED'
 export const PATIENT_REMOVE_PARENT_FAILED = 'PATIENT_REMOVE_PARENT_FAILED';
 export const PATIENT_REMOVE_PARENT_ACTION_STATUS = 'PATIENT_REMOVE_PARENT_ACTION_STATUS';
 
-export const PATIENT_AUTOCOMPLETE_REQUESTED = 'PATIENT_AUTOCOMPLETE_REQUESTED';
-export const PATIENT_AUTOCOMPLETE_SELECTED = 'PATIENT_AUTOCOMPLETE_SELECTED';
-export const PATIENT_AUTOCOMPLETE_SUCCEEDED = 'PATIENT_AUTOCOMPLETE_SUCCEEDED';
-export const PATIENT_AUTOCOMPLETE_FAILED = 'PATIENT_AUTOCOMPLETE_FAILED';
-
 export const PATIENT_SEARCH_REQUESTED = 'PATIENT_SEARCH_REQUESTED';
-export const PATIENT_SEARCH_SUCCEEDED = 'PATIENT_SEARCH_SUCCEEDED';
-export const PATIENT_SEARCH_FAILED = 'PATIENT_SEARCH_FAILED';
-
-export const CHANGE_SEARCH_TYPE_REQUESTED = 'CHANGE_SEARCH_TYPE_REQUESTED';
 
 export const PATIENT_SUBMISSION_LOCAL_PRACTITIONER = 'PATIENT_SUBMISSION_LOCAL_PRACTITIONER';
 export const PATIENT_SUBMISSION_LOCAL_SUPERVISOR = 'PATIENT_SUBMISSION_LOCAL_SUPERVISOR';
@@ -168,11 +143,6 @@ export const PATIENT_SUBMISSION_SERVICE_REQUEST_CHANGE_STATUS_SUCCEEDED =
   'PATIENT_SUBMISSION_SERVICE_REQUEST_CHANGE_STATUS_SUCCEEDED';
 export const PATIENT_SUBMISSION_SERVICE_REQUEST_CHANGE_STATUS_FAILED =
   'PATIENT_SUBMISSION_SERVICE_REQUEST_CHANGE_STATUS_FAILED';
-
-export const NANUQ_EXPORT_REQUESTED = 'NANUQ_EXPORT_REQUESTED';
-export const NANUQ_EXPORT_SUCCEEDED = 'NANUQ_EXPORT_SUCCEEDED';
-export const NANUQ_EXPORT_INVALID = 'NANUQ_EXPORT_INVALID';
-export const NANUQ_EXPORT_FAILED = 'NANUQ_EXPORT_FAILED';
 
 export const CLEAR_PATIENT_DATA_REQUESTED = 'CLEAR_PATIENT_DATA_REQUESTED';
