--- client/src/components/NanuqModal/styles.scss
@@ -1,7 +1,0 @@
-.nanuq-modal__content {
-  ul {
-    span.bold {
-      font-weight: bold;
-    }
-  }
-}
