--- client/src/components/screens/PatientSearch/components/style.scss
@@ -1,19 +1,0 @@
-.patients-table {
-  $icon-width: 16px;
-
-  &__cell-container {
-    display: flex;
-    height: 100% !important;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    margin-top: $icon-width / 2;
-  }
-
-  &__fetus-icon {
-    width: $icon-width;
-    margin-top: -$icon-width * 0.75;
-    margin-left: 4px;
-    fill: #1d8bc6;
-  }
-}
