+++ client/src/containers/AppRouter.jsx
@@ -15,7 +15,6 @@ import PropTypes from 'prop-types';
 
 import AccessDenied from 'components/screens/AccessDenied';
 import PatientScreen from 'components/screens/Patient';
-import PatientSearchScreen from 'components/screens/PatientSearch';
 import PatientsPrescriptionsSearch from 'components/screens/PatientsPrescriptionsSearch';
 import PatientSubmissionScreen from 'components/screens/PatientSubmission';
 import PatientVariantScreen from 'components/screens/PatientVariant';
@@ -81,15 +80,6 @@ const AppRouter = ({ history }) => {
           exact
           key="route-patient-search-local"
           path={`${Routes.PatientSearchArranger}/local`}
-          resource={KEYCLOAK_AUTH_RESOURCE_PATIENT_LIST}
-          roles={[]}
-        />
-
-        <AuthRoute
-          Component={PatientSearchScreen}
-          exact
-          key="route-patient-search-old"
-          path={Routes.PatientSearchOld}
           resource={KEYCLOAK_AUTH_RESOURCE_PATIENT_LIST}
           roles={[]}
         />
