--- client/src/helpers/search_table_helper.ts
@@ -1,10 +1,0 @@
-export const PATIENT_SEARCH_STORAGE_KEY = 'screen.patientsearch.table';
-
-export const defaultUserSearchColumns = [
-];
-
-export const defaultUserSearchColumnsOrder = [
-];
-
-export const LOCAL_STORAGE_PATIENT_SEARCH_COLUMNS_KEY = 'patient-search-columns';
-export const LOCAL_STORAGE_PATIENT_SEARCH_COLUMNS_ORDER_KEY = 'patient-search-columns-order';
