+++ client/src/reducers/index.ts
@@ -8,7 +8,6 @@ import patientEditionReducer, { PatientEditionState } from './patientEdition';
 import patientSubmissionReducer from './patientSubmission';
 import prescriptions, { PatientRequestCreationState } from './prescriptions';
 import prescriptionsGraphql, { PrescriptionState } from './prescriptionsGraphql';
-import searchReducer from './search';
 import serviceRequestCodeReducer, { ServiceRequestCodeState } from './serviceRequestCode';
 import userReducer from './user';
 
@@ -21,7 +20,6 @@ const rootReducer = (history: any) => combineReducers({
   prescriptions,
   prescriptionsGraphql,
   router: connectRouter(history),
-  search: searchReducer,
   serviceRequestCode: serviceRequestCodeReducer,
   user: userReducer,
 });
@@ -33,7 +31,6 @@ export type State = {
   app: any,
   user: any,
   patient: PatientState,
-  search: any,
   patientSubmission: any,
   patientCreation: PatientCreationState,
   patientEdition: PatientEditionState,
