+++ client/src/sagas/user.js
@@ -66,15 +66,6 @@ function* updateUserProfile(action) {
   }
 }
 
-function* columnsReset() {
-  try {
-    yield put({ type: actions.USER_PROFILE_UPDATE_COLUMNS_RESET });
-    yield put({ type: actions.USER_PROFILE_UPDATE_COLUMNS_RESET_SUCCEEDED });
-  } catch (e) {
-    yield put({ type: actions.USER_PROFILE_UPDATE_COLUMNS_RESET_FAILED });
-  }
-}
-
 function* updateUserAuthPermissions() {
   try {
     const response = yield Api.getUserAuthPermissions();
@@ -107,10 +98,6 @@ function* watchUpdateUserProfile() {
   yield takeLatest(actions.USER_PROFILE_UPDATE_REQUESTED, updateUserProfile);
 }
 
-function* watchColumnsReset() {
-  yield takeLatest(actions.USER_PROFILE_UPDATE_COLUMNS_RESET_REQUESTED, columnsReset);
-}
-
 function* watchUpdateUserAuthPermissions() {
   yield takeLatest(actions.UPDATE_USER_AUTH_PERMISSIONS_REQUESTED, updateUserAuthPermissions);
 }
@@ -121,7 +108,6 @@ export default function* watchedUserSagas() {
     watchUserLogout(),
     watchGetUserProfile(),
     watchUpdateUserProfile(),
-    watchColumnsReset(),
     watchUpdateUserAuthPermissions(),
   ]);
 }
